requirements.txt
jupyter==1.0.0
numpy==1.23.4
pandas==1.5.1
psycopg2-binary==2.9.5
sqlalchemy==1.4.44
alembic==1.8.1